+++ funcionalidades.h
@@ -36,7 +36,7 @@ INTEGRANTES DO GRUPO:
     void buscaIdArvore(int id, FILE *file, char *nomeArquivoArvoreB, int i, int buscaMinuscula);
 
     // Funcionalidade 10
-    void inserirArvoreB(FILE *arquivo, int chave, long long int byteOffset);
+    void inserirArvoreB(FILE *arquivo, int chave, int64_t byteOffset);
     void insercaoArvoreBRecursiva(FILE *arquivo, CABECALHO_ARVORE_B *cabecalho, int chave, int byteOffset, int rrnAtual, REGISTRO_ARVORE_B **caminho, int nivel, int *tamCaminho);
 
 #endif
