+++ funcoesAuxiliares.h
@@ -58,13 +58,13 @@ INTEGRANTES DO GRUPO:
     // Funções auxiliares do cabeçalho árvore-B
     CABECALHO *criarCabecalho(void);
     char getStatus(CABECALHO *cabecalho);
-    long long int getTopo(CABECALHO *cabecalho);
-    long long int getProxByteOffset(CABECALHO *cabecalho);
+    int64_t getTopo(CABECALHO *cabecalho);
+    int64_t getProxByteOffset(CABECALHO *cabecalho);
     int getNroRegArq(CABECALHO *cabecalho);
     int getNroRem(CABECALHO *cabecalho);
     void setStatus(CABECALHO *cabecalho, char status);
-    void setTopo(CABECALHO *cabecalho, long long int topo);
-    void setProxByteOffset(CABECALHO *cabecalho, long long int proxByteOffset);
+    void setTopo(CABECALHO *cabecalho, int64_t topo);
+    void setProxByteOffset(CABECALHO *cabecalho, int64_t proxByteOffset);
     void setNroRegArq(CABECALHO *cabecalho, int nroRegArq);
     void setNroRem(CABECALHO *cabecalho, int nroRem);
     void limpaCabecalho(CABECALHO *cabecalho);
@@ -89,17 +89,17 @@ INTEGRANTES DO GRUPO:
 
     // Funções auxiliares manipulação de registros árvore-B
     REGISTRO_ARVORE_B *criarRegistroArvoreBVazio();
-    int inserirChaveRegistroArvoreB(REGISTRO_ARVORE_B *registro, int chave, long long int byteOffset);
+    int inserirChaveRegistroArvoreB(REGISTRO_ARVORE_B *registro, int chave, int64_t byteOffset);
     int removerChaveRegistroArvoreB(REGISTRO_ARVORE_B *registro, int chave);
-    int inserirDescendenteRegistroArvoreB(REGISTRO_ARVORE_B *registro, long long int descendente, int chaveDescendente);
-    int removerDescendenteRegistroArvoreB(REGISTRO_ARVORE_B *registro, long long int descendente);
+    int inserirDescendenteRegistroArvoreB(REGISTRO_ARVORE_B *registro, int64_t descendente, int chaveDescendente);
+    int removerDescendenteRegistroArvoreB(REGISTRO_ARVORE_B *registro, int64_t descendente);
     int getRRNRegistroArvoreB(REGISTRO_ARVORE_B *registro);
     int getAlturaNoRegistroArvoreB(REGISTRO_ARVORE_B *registro);
     int setRRNRegistroArvoreB(REGISTRO_ARVORE_B *registro, int rrn);
     int setAlturaNoRegistroArvoreB(REGISTRO_ARVORE_B *registro, int alturaNo);
     int getNroChavesRegistroArvoreB(REGISTRO_ARVORE_B *registro);
     int getChave(REGISTRO_ARVORE_B *registro, int posicao);
-    long long int getByteOffsetRegistroArvoreB(REGISTRO_ARVORE_B *registro, int posicao);
+    int64_t getByteOffsetRegistroArvoreB(REGISTRO_ARVORE_B *registro, int posicao);
     int getDescendente(REGISTRO_ARVORE_B *registro, int posicao);
     int apagarRegistroArvoreB(REGISTRO_ARVORE_B *registro);
     REGISTRO_ARVORE_B *lerRegistroArvoreB(FILE *arquivo, int rrn);
@@ -112,7 +112,7 @@ INTEGRANTES DO GRUPO:
     DADOS* criarRegistro(char removido, int tamanhoRegistro, long prox, int id, int idade, int tamNomeJogador, char *nomeJogador, int tamNacionalidade, char *nacionalidade, int tamNomeClube, char *nomeClube);
     char retornaRemovido(DADOS *registro);
     int retornaTamanhoRegistro(DADOS *registro);
-    long long int retornaProx(DADOS *registro);
+    int64_t retornaProx(DADOS *registro);
     int retornaId(DADOS *registro);
     int retornaIdade(DADOS *registro);
     int RetornaTamNomeJogador(DADOS *registro);
@@ -122,7 +122,7 @@ INTEGRANTES DO GRUPO:
     char* retornaNacionalidade(DADOS *registro);
     int retornaTamNomeClube(DADOS *registro);
     char* retornaNomeClube(DADOS *registro);
-    void setaProx(DADOS *registro, long long int prox);
+    void setaProx(DADOS *registro, int64_t prox);
     void setaRemovido(DADOS *registro, int removido);
     void setaTamanhoRegistro(DADOS *registro, int tamanhoRegistro);
     void setaId(DADOS *registro, int id);
@@ -141,10 +141,10 @@ INTEGRANTES DO GRUPO:
     bool escreverRegistro(DADOS *registro, int byteOffset, int tamRegistroAtual, FILE *arquivoBin);
     void imprimeRegistrosBuscados(char *arquivo, int buscaId, char *nomeArquivoArvoreB);
     void imprimirRegistrosPorCampos(FILE *file, CABECALHO *cabecalho, int buscaId, char *nomeArquivoArvoreB, int i);
-    long long int buscarRegistroIdRec(FILE *fileArvoreB, int id, int rrnAtual);
+    int64_t buscarRegistroIdRec(FILE *fileArvoreB, int id, int rrnAtual);
     REMOVIDOS *criarListaRemovidos(FILE *file);
     bool inserirNovoDadoArvoreB(char *arquivoBinario, char *arquivoArvoreB, int numOperacoes);
-    bool adicionarNoArvoreB(int chave, long long int byteOffset, FILE *arquivoArvoreB);
+    bool adicionarNoArvoreB(int chave, int64_t byteOffset, FILE *arquivoArvoreB);
     REMOVIDOS *criarListaRemovidosVazia();
     void apagarListaRemovidos(REMOVIDOS *removidos);
     void removerRegistroIndice(LISTA_INDICE *lista, int index);
@@ -152,7 +152,7 @@ INTEGRANTES DO GRUPO:
     REGISTRO_INDICE *criarRegistroIndice();
     void removerRegistroRemovidoEAtualizarArquivo(REMOVIDOS *removidos, int posicao, FILE *file);
     void setIndexRegistroIndice(REGISTRO_INDICE *registro, int index);
-    void setByteOffsetRegistroIndice(REGISTRO_INDICE *registro, long long int byteOffset);
+    void setByteOffsetRegistroIndice(REGISTRO_INDICE *registro, int64_t byteOffset);
     void setRegistroListaIndice(LISTA_INDICE *lista, int index, REGISTRO_INDICE *registro);
     void setTamanho(LISTA_INDICE *lista, int tamanho);
     LISTA_INDICE *criarListaIndice();
@@ -160,13 +160,13 @@ INTEGRANTES DO GRUPO:
     void apagarRegistroIndice(REGISTRO_INDICE *registro);
     void adicionarRegistroRemovido(REMOVIDOS *removidos, REGISTRO_INDICE *registroIndice, int tamanho);
     int getTamanhoListaIndice(LISTA_INDICE *lista);
-    long long int getBestFitAndFreeSpace(REMOVIDOS *removidos, int tamanho, DADOS *registro, FILE *file);
-    long long int getMaiorByteOffsetMenorQue(REMOVIDOS *removidos, int id);
+    int64_t getBestFitAndFreeSpace(REMOVIDOS *removidos, int tamanho, DADOS *registro, FILE *file);
+    int64_t getMaiorByteOffsetMenorQue(REMOVIDOS *removidos, int id);
     void shiftElementosListaRemovidosRight(REMOVIDOS *removidos, int pos);
     REGISTRO_INDICE *getRegistroIndice(LISTA_INDICE *lista, int index);
     int buscarPosicaoRegistroIndiceLinear(LISTA_INDICE *lista, int id);
     int getIndexRegistroIndice(REGISTRO_INDICE *registro);
-    long long int getByteOffsetRegistroIndice(REGISTRO_INDICE *registro);
+    int64_t getByteOffsetRegistroIndice(REGISTRO_INDICE *registro);
     int getTamanhoListaRemovidos(REMOVIDOS *removidos);
 
 #endif
